Changed data storage format, finished scripts
