+++ server/src/data/01_hex_variables_non_gov_regardless_of_reporting_frequency_20210209.csv
@@ -1,4 +1,4 @@
-Index	HEX_600	01_HEX_weighted_average_distance	01_HEX_weighted_mean_home_distance	01_HEX_weighted_max_home_distance	01_HEX_weighted_median_home_distance	01_HEX_weighted_radius_of_gyration	01_HEX_weighted_num_reporting_2015	01_HEX_weighted_HEX_coverage	01_HEX_weighted_subject_coverage	01_HEX_weighted_subject_hhi	01_HEX_total_reporting	01_HEX_total_user	user_type	reporting_level
+	HEX_600	01_HEX_weighted_average_distance	01_HEX_weighted_mean_home_distance	01_HEX_weighted_max_home_distance	01_HEX_weighted_median_home_distance	01_HEX_weighted_radius_of_gyration	01_HEX_weighted_num_reporting_2015	01_HEX_weighted_HEX_coverage	01_HEX_weighted_subject_coverage	01_HEX_weighted_subject_hhi	01_HEX_total_reporting	01_HEX_total_user	user_type	reporting_level
 0	8	2074.2733301211	1995.7984093508	5153.7309571418	1810.7429082696	758.4117628274	30	23.3333333333	2.6666666667	0.7584279791	36	18	1	1
 1	9	166.6819314381	404.4524680004	514.6821331557	391.9728319735	23.7505939393	1.8235294118	1.068627451	1.1862745098	0.8771786492	10404	7752	1	1
 2	10	287.6614099718	257.5782153472	571.1128273883	208.70038488	30.8916415974	2.3950617284	1.1604938272	1.2098765432	0.8952161194	6561	4536	1	1
