+++ server/src/scripts/csvToMongo.py
@@ -1,42 +1,29 @@
-#!/usr/bin/env python3
 import os
-import sys
 import pandas as pd
-import pymongo
+from pymongo import MongoClient
 import json
 
+def decorateRow(row):
+    #['', 'HEX_600', 'HEX_weighted_average_distance', 'HEX_weighted_mean_home_distance', 'HEX_weighted_max_home_distance', 'HEX_weighted_median_home_distance', 'HEX_weighted_radius_of_gyration', 'HEX_weighted_num_reporting_2015', 'HEX_weighted_HEX_coverage', 'HEX_weighted_subject_coverage', 'HEX_weighted_subject_hhi', 'HEX_total_reporting', 'HEX_total_user']
+    record = row[2:].to_frame().transpose()
+    return record
 
+client = MongoClient('localhost', 27017)
+db = client['hexData']
+coll = db['hexagon']
 
-def import_content(filepath):
-    mng_client = pymongo.MongoClient('127.0.0.1', 27017)
-    mng_db = mng_client['hexData'] 
-    collection_name = 'hexagon' 
-    db_cm = mng_db[collection_name]
-    cdir = os.path.dirname(__file__)
-    file_res = os.path.join(cdir, filepath)
-
-    data = pd.read_csv(file_res)
-    data_json = json.loads(data.to_json(orient='records'))
-
-
-    db_cm.insert_many(data_json)
-
-def decorate_metadata(filepath, data_json):
-    fileNum = os.path.basename(filepath)[0:2]
-
-    data_json["non_gov_users"][0:] = True
-    data_json["unsure_user"][0:] = False
-    data_json["reported_more_than_2x"][0:] = False
-
-    if (fileNum == "02" or fileNum == "03" or fileNum == "05" or fileNum == "06"):
-        data_json["unsure_user"][0:] = True
-    if (fileNum == "04" or fileNum == "05" or fileNum == "06"):
-        data_json["unsure"][0:] = True
-    if (    
-
-
-
-
-if __name__ == "__main__":
-    filepath = '/vagrant/src/data/01_hex_variables_non_gov_regardless_of_reporting_frequency_20210209.csv'
-import_content(filepath)
+directory = "/vagrant/src/data"
+for filename in os.listdir(directory):
+    if filename.endswith(".csv"):
+        data = pd.read_csv(os.path.join(directory, filename), sep='\t')
+        for index, row in data.iterrows():
+            # print(json.loads(decorateRow(row).to_json(orient='records'))[0])
+            coll.update_one({
+                'HEX_600': row.HEX_600,
+            },
+            {
+                '$push': 
+                {
+                    'results': json.loads(decorateRow(row).to_json(orient='records'))[0]
+                }
+            })
