+++ server/src/scripts/dump/admin/system.version.metadata.json
@@ -1,0 +1,1 @@
+{"indexes":[{"v":{"$numberInt":"2"},"key":{"_id":{"$numberInt":"1"}},"name":"_id_"}],"uuid":"3f6a2043cfc54483b52cd192370e1274","collectionName":"system.version","type":"collection"}
