+++ server/src/scripts/dump/hexData/hexagon.metadata.json
@@ -1,0 +1,1 @@
+{"indexes":[{"v":{"$numberInt":"2"},"key":{"_id":{"$numberInt":"1"}},"name":"_id_"}],"uuid":"9fa2776ff6dd4c67955538142e4b7790","collectionName":"hexagon","type":"collection"}
