+++ server/src/scripts/dump/hexData/hexagons.metadata.json
@@ -1,0 +1,1 @@
+{"indexes":[{"v":{"$numberInt":"2"},"key":{"_id":{"$numberInt":"1"}},"name":"_id_"}],"uuid":"40a5dc0a789d463baa2b358223d6f3e8","collectionName":"hexagons","type":"collection"}
